python-backend/requirements.txt
# FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Database
motor>=3.0.0
pymongo>=4.0.0
beanie>=1.20.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.0
bcrypt>=4.0.0

# AI/ML
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
httpx>=0.24.0
email-validator>=2.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
